split dtype 32 to two files
